+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "graph-canonicalization"
+name = "graph_canonicalization"
 version = "0.1.0"
 authors = ["Qin Wan <rogerwq@gmail.com>"]
 edition = "2018"
