--- src/ext/molecule/tests/krotko.md
@@ -1,39 +1,0 @@
-# Comments on the paper [Atomic ring invariant and Modified CANON extended connectivity algorithm for symmetry perception in molecular graphs and rigorous canonicalization of SMILES](https://jcheminf.biomedcentral.com/articles/10.1186/s13321-020-00453-4#Sec13)
-
-_ATTENTION: This paper uses abbrevations from Reference 1 authored by Ivanciuc, the GIAP and CCAP, meanwhile in our paper we use GIVP and CNAP instead._
-
-## The understanding
-
-For the GIVP part, this paper follows the work by [Nadine Schneider](https://pubs.acs.org/doi/abs/10.1021/acs.jcim.5b00543) by proposing another two atom invariants: ring invariant and distance invariant. As a member of the family of the graph relaxation algorithms, it improves the CANON algorithm from [Weininger](https://pubs.acs.org/doi/pdf/10.1021/ci00062a008) too. Therefore, the GIVP part is composed of generation of atom invariants and the improved CANON algorithm.
-
-The paper makes an effort on generation of SMILES string. Based on the partitioning results from the GIVP part, a "break tie" process is carried on for each ring atom, therefore a consequential SMILES string. The SMILES strings are comparable, thus one of them can be selected as the canonical SMILES string.
-
-
-## Questions
-
-### Does the modified CANON algorithm completely solve the oscillary behavoirs of graph relaxation?
-
-The oscillary behavoirs of graph relaxation depends on the initial value of numbering (or ranking) and the topological connectivies. Is it possible to prove that the modified CANON algorithm can get rid of the oscillary behavoirs?
-
-### Is CNAP for symmetry perception implemented?
-
-As stated by the paper, "This means that regardless of the accuracy of the symmetry determination" in Page 10, the answer shall be no. The canonical SMILES generated by the algorithm in this paper could be rigorous, but the symmetrical groups are still remained to be investigated.
-
-## Comparison by examples
-
-I have made an experiment on some examples from this paper by using our approach.
-
-### C12C3C1C3C1C3C2C13
-This is the example of figure 2. Our approach gives the right answer of 2 symmetrical classes with no specific atom invariant needed.
-
-### C1(C2C3C4C15)C6C7C2C8C3C9C%10C4C%11C5C6C%12C%11C%10C%13C%12C7C8C9%13
-This is the example of figure 3. The GIVP result of our approach gives the right answer of 2 symmetrical classes with no specific atom invariant needed. The CNAP process failed since this is a highly symmetrical graph and the computation required is beyond the limit of the setting.
-
-### NC1=CC=C1
-This is the example of figure 6. Without graph relaxation, our approach gives the right answer.
-
-### C12C3C4C5C1C6C7C2C8C3C6C5C8C74
-This is an example from the discussion section. It is a highly symmetrical graph and our approach gives the right answer. The CNAP process succeeds too.
-
-### C1OC23COC45COC11COC67COC8(COC9(CO2)COC(CO1)(CO6)OCC(CO9)(OC4)OCC(CO5)(OC7)OC8)OC3
-This is a failure case of the algorithm, and the Schneider algorithms failed on this example too. However, our approach gives the right answer.
