--- src/ext/molecule/tests/mod.rs
@@ -1,2 +1,0 @@
-mod chembl;
-mod krotko;
